+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * ENHANCEMENT: AGI will not parse and graph `security.c` type messages for authentication message counts per second.
 * ENHANCEMENT: WebUI add option to `attach trino` for the trino client cli.
 * ENHANCEMENT: GCP Expiry system will attempt to enable required services prior to installation.
+* ENHANCEMENT: Docker: for version 7+ of Aerospike, if a custom config file is not provided, modify the default one to remove `bar` namespace and change `test` to use file-backed `storage-engine device`.
 * FIX: AGI: Ingest `aggregate` would incorrectly miss aggregation values when working with lines that contain `(repeated:X)`.
 
 #### 7.4.0