+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * FIX: AGI: Handle connection error during installation gracefully.
 * FIX: WebUI: make listener bind specifically to IPv4 or IPv6 depending on specified address instead of letting the kernel decide.
 * FIX: WebUI: do not clear running jobs from history in the web interface until they are completed (when clear history button is pressed).
+* FIX: GCP: For accelerated instances fallback to `onHostMaintenance=TERMINATE` policy.
 
 #### 7.4.0
 * FEATURE: Web UI.